dotnet-sdk-setup-hook: fix case where ln fails because file exists

+++ pkgs/development/compilers/dotnet/dotnet-sdk-setup-hook.sh
@@ -23,7 +23,9 @@ _linkPackages() {
         for x in "$src"/*/*; do
             dir=$dest/$(basename "$(dirname "$x")")
             mkdir -p "$dir"
-            ln -s "$x" "$dir"/
+            if [[ ! -f "$dir/$(basename "$x")" ]]; then
+              ln -s "$x" "$dir"/
+            fi
         done
     )
 }
